Add files via upload
Description: This project is focused on analyzing the risk associated with a financial portfolio. It calculates the historical log returns of the assets in the portfolio and computes the covariance and correlation matrices. Using these, the project estimates the Value at Risk (VaR) and Conditional Value at Risk (CVaR) for both the portfolio as a whole and for each individual asset. Additionally, it calculates the beta and marginal VaR for each asset, providing insights into their respective contributions to the portfolio's overall risk.

Analysis of Results: The output provides a comprehensive view of the risk profile of the portfolio. The VaR and CVaR values give an estimate of the potential losses that could be incurred. The beta values indicate the sensitivity of each asset to market movements, and the marginal VaR highlights the incremental risk each asset adds to the portfolio. By analyzing these results, investors can make informed decisions about asset allocation and risk management in their portfolio.
Portfolio Summary:

The total CVaR of the portfolio is $7,403.29, indicating the potential loss in value at a 98% confidence level over the specified time period.
The assets with the highest contribution to the portfolio's risk are MCD, JPM, and WMT, as indicated by their higher CVaR percentages and beta values.
The assets with the lowest contribution to the portfolio's risk are KO and AAPL, as indicated by their lower CVaR percentages and beta values.
The portfolio's overall risk is diversified, with different assets contributing varying degrees of risk. This diversification helps in managing the overall risk exposure of the portfolio.